pyproject.toml
[project]
name = "taam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "qrcode>=8.2",
    "uvicorn>=0.35.0",
]
